requirements.txt
# AWS SDK
boto3==1.34.0
botocore==1.34.0

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0